Agrupada en las rutas api con sus midleware las demas rutas que faltaban del commit anterior, ademas se agrego un nuevo template en el component admin el cual cambia la vista y me termino gustando mas luego la navbar sera cambiada al estilo de este html con su  css, se agrego en servicios.service.ts con el header donde esta la autentificacion del usuario, el cual hace las peticiones a los endpoint

+++ MascotasFresh-Backend/routes/api.php
@@ -18,24 +18,27 @@ use Illuminate\Support\Facades\Route;
     return $request->user();
 });*/
 Route::post('login', 'ApiLoginController@login');
+Route::get('/user', 'UserController@user')->middleware('auth:api');
 /*----------------Rutas Middleware---------------------------*/
 Route::middleware(['auth:api','rol'])->group(function() {
+    /*--------------- Rutas Animales ----------------------------*/
     Route::middleware(['scope:Admin,Medico'])->get('/animales','AnimalController@index');
-    
+    Route::middleware(['scope:Admin,Medico'])->post('/agregaranimal', 'AnimalController@store');
+    Route::middleware(['scope:Admin,Medico'])->put('/modificaranimal', 'AnimalController@update');
+    Route::middleware(['scope:Admin,Medico'])->delete('/eliminaranimal/{id}', 'AnimalController@destroy');
+
+    /*----------------------------citas-------------------------------*/
+    Route::middleware(['scope:Admin,Medico'])->post('/agregarcita', 'MedicoServicioAnimalController@store');
+    Route::middleware(['scope:Admin,Medico'])->get('/serviciosDisponibles','ServicioController@index');
+
+    /*----------------Rutas Usuario ---------------------------------*/
+    //Route::post('/registro', 'UserController@register'); no usaremos el registro hasta que existan algunos usuarios fuera de la veterinaria
+    Route::middleware(['scope:Admin'])->get('/users','UserController@index');
+    Route::middleware(['scope:Admin'])->post('/newuser','UserController@store');
+
 });
-// lo que pasa es lo que explican en laracast que el token se crea al pedir la peticion y ese metodo de checkrole tengo que hacerlo en la aplicacion 
+// lo que pasa es lo que explican en laracast que el token se crea al pedir la peticion y ese metodo de checkrole tengo que hacerlo en la aplicacion
 
-/*--------------- Rutas Animales ----------------------------*/
-Route::post('/agregaranimal', 'AnimalController@store');
-Route::put('/modificaranimal', 'AnimalController@update');
-Route::delete('/eliminaranimal/{id}', 'AnimalController@destroy');
-/*----------------Rutas Usuario ---------------------------------*/
 
-//Route::post('/registro', 'UserController@register'); no usaremos el registro hasta que existan algunos usuarios fuera de la veterinaria
-Route::get('/users','UserController@index');
-Route::post('/newuser','UserController@store');
 
-/*----------------------------citas-------------------------------*/
-Route::post('/agregarcita', 'MedicoServicioAnimalController@store');
-Route::get('/serviciosDisponibles','ServicioController@index');
